chatbot/requirements.txt
langgraph
langchain
langchain-google-genai
langchain-groq
langchain-tavily
langchain-community
langchain-huggingface
langchain-pinecone
pypdf
PyPDF2
langchain_community
fastapi
uvicorn
pydantic
python-dotenv
streamlit